1stapril2020

+++ screens/UserProductsScreen.js
@@ -1,10 +1,11 @@
-import React from "react";
-import { View, FlatList, TouchableOpacity } from "react-native";
+import React, { useState, useCallback } from "react";
+import { View, FlatList, TouchableOpacity, Alert } from "react-native";
 import { Icon } from "react-native-elements";
 import { HeaderButtons, Item } from "react-navigation-header-buttons";
 import HeaderButton from "../components/UI/HeaderButton";
-
+import { Snackbar } from "react-native-paper";
 import * as productActions from "../store/actions/product";
+import Colors from "../constants/Colors";
 
 import Product from "../components/shop/Product";
 import { useSelector, useDispatch } from "react-redux";
@@ -14,7 +15,30 @@ const UserProductsScreen = props => {
 
   const dispatch = useDispatch();
 
+  const [isVisible, setIsVisible] = useState(false);
+  const [listener, setListener] = useState(false);
+  const [error, setError] = useState();
+
   const renderUserProductItem = itemData => {
+    // const deleteconfirmHandler = () => {
+    //   dispatch(productActions.deleteUserProduct(itemData.item.id));
+    //   setIsVisible(true);
+    // };
+    // const deleteconfirmHandler = useCallback(async () => {
+    //   // console.log("LOAD");
+
+    //   setError(null);
+    //   try {
+    //     await dispatch(productActions.deleteUserProduct(itemData.item.id));
+    //   } catch (err) {
+    //     setError(err.message);
+    //   }
+    //   setIsVisible(true);
+    // }, [setIsVisible, setError, dispatch]);
+
+    // const deleteconfirmHandlerWrapper = () => {
+    //   deleteconfirmHandler();
+    // };
     return (
       <Product
         image={itemData.item.imageUrl}
@@ -32,7 +56,25 @@ const UserProductsScreen = props => {
       >
         <TouchableOpacity
           onPress={() => {
-            dispatch(productActions.deleteUserProduct(itemData.item.id));
+            Alert.alert("Delete", "Do you really want to dewlete it oppa ?", [
+              {
+                text: "delete",
+                style: "destructive",
+                onPress: () => {
+                  dispatch(productActions.deleteUserProduct(itemData.item.id));
+                }
+              },
+              {
+                text: "cancel",
+                style: "cancel",
+                onPress: () => {
+                  setListener(true);
+                }
+              }
+            ]);
+            if (listener) {
+              setIsVisible(true);
+            }
           }}
         >
           <Icon
@@ -49,13 +91,25 @@ const UserProductsScreen = props => {
   };
 
   return (
-    <View>
+    <View style={{ flex: 1 }}>
       <FlatList
         keyExtractor={item => item.id}
         data={userProducts}
         renderItem={renderUserProductItem}
         contentContainerStyle={{ marginHorizontal: 10, flexGrow: 1 }}
       />
+      <Snackbar
+        visible={isVisible}
+        onDismiss={() => setIsVisible(false)}
+        duration={2000}
+        action={{
+          label: "UNDO",
+          onPress: () => {}
+        }}
+        style={{ backgroundColor: Colors.primaryColor, height: 50 }}
+      >
+        Item Deleted
+      </Snackbar>
     </View>
   );
 };
